+++ src/config/index.ts
@@ -11,6 +11,7 @@
  * @packageDocumentation
  */
 
+import { RequestErrorDetailsExtractor } from 'core/request/error';
 import type { Config } from 'config/interface';
 
 export * from 'config/interface';
@@ -33,7 +34,11 @@ const config = <Config>{
 	log: {
 		pipelines: [
 			{
-				middlewares: ['configurable'],
+				middlewares: [
+					'configurable',
+					['extractor', [new RequestErrorDetailsExtractor()]]
+				],
+
 				engine: 'console',
 				engineOptions: {
 					default: {