+++ src/core/log/middlewares/extractor/const.ts
@@ -1,0 +1,12 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+/**
+ * Maximum depth of nested errors
+ */
+export const DEPTH_LIMIT = 10;