+++ src/core/log/middlewares/extractor/index.ts
@@ -1,0 +1,15 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+/**
+ * [[include:core/log/middlewares/extractor/README.md]]
+ * @packageDocumentation
+ */
+
+export { ExtractorMiddleware } from 'core/log/middlewares/extractor/extractor';
+export * from 'core/log/middlewares/extractor/interface';