+++ src/core/request/error/interface.ts
@@ -6,14 +6,18 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
-//#if node_js
 import type Response from 'core/request/response';
-//#endif
-
 import type { NormalizedCreateRequestOptions } from 'core/request/interface';
 
 export interface Details<D = unknown> extends Dictionary {
 	request?: NormalizedCreateRequestOptions<D>;
 	response?: Response<D>;
 	error?: object;
+}
+
+export interface RequestErrorDetailsExtractorSettings {
+	headers: {
+		include?: string[];
+		exclude?: string[];
+	};
 }