+++ src/core/log/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.37.0 (2021-03-17)
+
+#### :boom: Breaking Change
+
+* Now `details` property within the `LogEvent` object is deprecated.
+  The `additionals` property should be using instead.
+
+#### :rocket: New Feature
+
+* Added ability to set up middlewares via the config
+* Added a new middleware `ExtractorMiddleware`
+
 ## v3.29.0 (2020-12-22)
 
 #### :house: Internal